installer: add package verification to CI

+++ dockerfiles/verify_packages/verify_rpm_centos6.sh
@@ -10,10 +10,19 @@ fi
 for phpVer in $(ls ${PHP_INSTALL_DIR}); do
     echo "Installing ddtrace on PHP ${phpVer}..."
     switch-php $phpVer
-    rpm -Uvh /build_src/build/packages/*.rpm
-    php --ri=ddtrace
 
-    # Uninstall the tracer
-    rpm -e datadog-php-tracer
-    rm -f /opt/datadog-php/etc/ddtrace.ini
+    # Installing dd-trace-php
+    install_type="${install_type:-php_installer}"
+    if [ "$install_type" = "native_package" ]; then
+        echo "Installing dd-trace-php using the OS-specific package installer"
+        rpm -Uvh /build_src/build/packages/*.rpm
+        php --ri=ddtrace
+
+        # Uninstall the tracer
+        rpm -e datadog-php-tracer
+        rm -f /opt/datadog-php/etc/ddtrace.ini
+    else
+        echo "Installing dd-trace-php using the new PHP installer"
+        php dd-php-setup.php --tracer-file="$(pwd)/build/packages/*.tar.gz" --php-bin=all
+    fi
 done
